requirements.txt
asgiref==3.6.0
Django==4.1.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
greenlet==2.0.2
msgpack==1.0.4
mysqlclient==2.1.1
neovim==0.3.1
pynvim==0.4.3
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
